// src/app/models/user.ts
export interface UserModel {
  id: number;
  name: string;
  lastName: string;
  username: string;
  email: string;
  phoneNumber: number;
  password: string;
}
export interface UsersApi {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}